Adapt to new widgets feature
