+++ src/Field/Part/Hint.php
@@ -7,6 +7,7 @@ namespace Yiisoft\Form\Field\Part;
 use InvalidArgumentException;
 use Stringable;
 use Yiisoft\Form\Field\Base\FormAttributeTrait;
+use Yiisoft\Form\ThemeDispatcher;
 use Yiisoft\Html\Html;
 use Yiisoft\Widget\Widget;
 
@@ -124,5 +125,10 @@ final class Hint extends Widget
             : Html::tag($this->tag, $content, $this->attributes)
                 ->encode($this->encode)
                 ->render();
+    }
+
+    protected static function getDefaultConfig(): array
+    {
+        return ThemeDispatcher::getTheme()->getHintConfig();
     }
 }
