+++ src/Field/Part/Label.php
@@ -6,6 +6,7 @@ namespace Yiisoft\Form\Field\Part;
 
 use Stringable;
 use Yiisoft\Form\Field\Base\FormAttributeTrait;
+use Yiisoft\Form\ThemeDispatcher;
 use Yiisoft\Html\Html;
 use Yiisoft\Widget\Widget;
 
@@ -149,5 +150,10 @@ final class Label extends Widget
         }
 
         return $tag->render();
+    }
+
+    protected static function getDefaultConfig(): array
+    {
+        return ThemeDispatcher::getTheme()->getLabelConfig();
     }
 }
